Separated views publishing into its own command
